+++ pyas2/models.py
@@ -1,32 +1,32 @@
 # -*- coding: utf-8 -*-
 import os
-import requests
 import traceback
+from email.parser import HeaderParser
+from uuid import uuid4
+
+import requests
 from django.core.files.base import ContentFile
 from django.db import models
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from django.utils import timezone
 from django.utils.translation import ugettext as _
-from email.parser import HeaderParser
-from pyas2lib import Mdn as As2Mdn
-from pyas2lib import Message as As2Message
-from pyas2lib import Organization as As2Organization
-from pyas2lib import Partner as As2Partner
 
-from uuid import uuid4
+from pyas2lib import (
+    Mdn as As2Mdn,
+    Message as As2Message,
+    Organization as As2Organization,
+    Partner as As2Partner
+)
+from pyas2lib.utils import extract_certificate_info
 
 from pyas2 import settings
 from pyas2.utils import run_post_send
 from pyas2.utils import store_file
-from pyas2lib.utils import extract_certificate_info
 
 
 class PrivateKey(models.Model):
     name = models.CharField(max_length=255)
     key = models.BinaryField()
-    key_pass = models.CharField(
-        max_length=100, verbose_name='Private Key Password')
+    key_pass = models.CharField( max_length=100, verbose_name='Private Key Password')
     valid_from = models.DateTimeField(null=True, blank=True)
     valid_to = models.DateTimeField(null=True, blank=True)
     serial_number = models.CharField(max_length=64, null=True, blank=True)
@@ -46,8 +46,7 @@ class PrivateKey(models.Model):
 class PublicCertificate(models.Model):
     name = models.CharField(max_length=255)
     certificate = models.BinaryField()
-    certificate_ca = models.BinaryField(
-        verbose_name=_('Local CA Store'), null=True, blank=True)
+    certificate_ca = models.BinaryField(verbose_name=_('Local CA Store'), null=True, blank=True)
     verify_cert = models.BooleanField(
         verbose_name=_('Verify Certificate'), default=True,
         help_text=_('Uncheck this option to disable certificate verification.'))
@@ -68,8 +67,7 @@ class PublicCertificate(models.Model):
 
 
 class Organization(models.Model):
-    name = models.CharField(
-        verbose_name=_('Organization Name'), max_length=100)
+    name = models.CharField(verbose_name=_('Organization Name'), max_length=100)
     as2_name = models.CharField(
         verbose_name=_('AS2 Identifier'), max_length=100, primary_key=True)
     email_address = models.EmailField(null=True, blank=True)
@@ -138,14 +136,11 @@ class Partner(models.Model):
         ('ASYNC', 'Asynchronous'),
     )
 
-    name = models.CharField(
-        verbose_name=_('Partner Name'), max_length=100)
-    as2_name = models.CharField(
-        verbose_name=_('AS2 Identifier'), max_length=100, primary_key=True)
+    name = models.CharField(verbose_name=_('Partner Name'), max_length=100)
+    as2_name = models.CharField(verbose_name=_('AS2 Identifier'), max_length=100, primary_key=True)
     email_address = models.EmailField(null=True, blank=True)
 
-    http_auth = models.BooleanField(
-        verbose_name=_('Enable Authentication'), default=False)
+    http_auth = models.BooleanField(verbose_name=_('Enable Authentication'), default=False)
     http_auth_user = models.CharField(max_length=100, null=True, blank=True)
     http_auth_pass = models.CharField(max_length=100, null=True, blank=True)
     https_verify_ssl = models.BooleanField(
@@ -153,15 +148,11 @@ class Partner(models.Model):
         help_text=_('Uncheck this option to disable SSL certificate verification to HTTPS.'))
 
     target_url = models.URLField()
-    subject = models.CharField(
-        max_length=255, default=_('EDI Message sent using pyas2'))
+    subject = models.CharField(max_length=255, default=_('EDI Message sent using pyas2'))
     content_type = models.CharField(
-        max_length=100, choices=CONTENT_TYPE_CHOICES,
-        default='application/edi-consent'
-    )
+        max_length=100, choices=CONTENT_TYPE_CHOICES, default='application/edi-consent')
 
-    compress = models.BooleanField(
-        verbose_name=_('Compress Message'), default=False)
+    compress = models.BooleanField(verbose_name=_('Compress Message'), default=False)
     encryption = models.CharField(
         max_length=20, verbose_name=_('Encrypt Message'),
         choices=ENCRYPT_ALG_CHOICES, null=True, blank=True)
@@ -176,8 +167,7 @@ class Partner(models.Model):
     )
 
     mdn = models.BooleanField(verbose_name=_('Request MDN'), default=False)
-    mdn_mode = models.CharField(
-        max_length=20, choices=MDN_TYPE_CHOICES, null=True, blank=True)
+    mdn_mode = models.CharField(max_length=20, choices=MDN_TYPE_CHOICES, null=True, blank=True)
     mdn_sign = models.CharField(
         max_length=20, verbose_name=_('Request Signed MDN'),
         choices=SIGN_ALG_CHOICES, null=True, blank=True)
@@ -195,8 +185,8 @@ class Partner(models.Model):
         verbose_name=_('Keep Original Filename'),
         default=False,
         help_text=_(
-            'Use Original Filename to to store file on receipt, use this option'
-            ' only if you are sure partner sends unique names')
+            'Use Original Filename to to store file on receipt, use this option '
+            'only if you are sure partner sends unique names')
     )
     cmd_send = models.TextField(
         verbose_name=_('Command on Message Send'),
@@ -234,15 +224,13 @@ class Partner(models.Model):
         if self.signature_cert:
             params['verify_cert'] = bytes(self.signature_cert.certificate)
             if self.signature_cert.certificate_ca:
-                params['verify_cert_ca'] = bytes(
-                    self.signature_cert.certificate_ca)
+                params['verify_cert_ca'] = bytes(self.signature_cert.certificate_ca)
             params['validate_certs'] = self.signature_cert.verify_cert
 
         if self.encryption_cert:
             params['encrypt_cert'] = bytes(self.encryption_cert.certificate)
             if self.encryption_cert.certificate_ca:
-                params['encrypt_cert_ca'] = bytes(
-                    self.encryption_cert.certificate_ca)
+                params['encrypt_cert_ca'] = bytes(self.encryption_cert.certificate_ca)
             params['validate_certs'] = self.encryption_cert.verify_cert
 
         if self.confirmation_message:
@@ -256,19 +244,16 @@ class Partner(models.Model):
 
 class MessageManager(models.Manager):
 
-    def create_from_as2message(self, as2message, payload, direction, status,
+    def create_from_as2message(self, as2message, payload, direction, status, filename=None,
                                detailed_status=None):
         """Create the Message from the pyas2lib's Message object"""
 
         if direction == 'IN':
-            organization = as2message.receiver.as2_name \
-                if as2message.receiver else None
+            organization = as2message.receiver.as2_name if as2message.receiver else None
             partner = as2message.sender.as2_name if as2message.sender else None
         else:
-            partner = as2message.receiver.as2_name \
-                if as2message.receiver else None
-            organization = as2message.sender.as2_name \
-                if as2message.sender else None
+            partner = as2message.receiver.as2_name if as2message.receiver else None
+            organization = as2message.sender.as2_name if as2message.sender else None
 
         message, _ = self.update_or_create(
             message_id=as2message.message_id,
@@ -285,34 +270,28 @@ class MessageManager(models.Manager):
         )
 
         # Save the headers and payload to store
-        message.headers.save(name='%s.header' % uuid4(),
-                             content=ContentFile(as2message.headers_str))
-        message.payload.save(name='%s.msg' % uuid4(),
-                             content=ContentFile(payload))
+        if not filename:
+            filename = f'{uuid4()}.msg'
+        message.headers.save(name=f'{filename}.header', content=ContentFile(as2message.headers_str))
+        message.payload.save(name=filename, content=ContentFile(payload))
 
         # Save the payload to the inbox folder
-        full_fn = None
+        full_filename = None
         if direction == 'IN' and status == 'S':
-            folder = os.path.join(
-                settings.DATA_DIR, 'messages', organization, 'inbox', partner)
-            if message.partner.keep_filename and \
-                    as2message.payload.get_filename():
-                filename = as2message.payload.get_filename()
-            else:
-                filename = '%s.msg' % message.message_id
-            full_fn = store_file(folder, filename, payload)
+            folder = os.path.join(settings.DATA_DIR, 'messages', organization, 'inbox', partner)
+            if not message.partner.keep_filename or not filename:
+                filename = f'{message.message_id}.msg'
+            full_filename = store_file(folder, filename, payload)
 
-        return message, full_fn
+        return message, full_filename
 
 
 def get_message_store(instance, filename):
     current_date = timezone.now().strftime('%Y%m%d')
     if instance.direction == 'OUT':
-        target_dir = os.path.join(
-            'messages', '__store', 'payload', 'sent', current_date)
+        target_dir = os.path.join('messages', '__store', 'payload', 'sent', current_date)
     else:
-        target_dir = os.path.join(
-            'messages', '__store', 'payload', 'received', current_date)
+        target_dir = os.path.join('messages', '__store', 'payload', 'received', current_date)
     return '{0}/{1}'.format(target_dir, filename)
 
 
@@ -340,15 +319,11 @@ class Message(models.Model):
     status = models.CharField(max_length=2, choices=STATUS_CHOICES)
     detailed_status = models.TextField(null=True)
 
-    organization = models.ForeignKey(
-        Organization, null=True, on_delete=models.SET_NULL)
-    partner = models.ForeignKey(
-        Partner, null=True, on_delete=models.SET_NULL)
+    organization = models.ForeignKey(Organization, null=True, on_delete=models.SET_NULL)
+    partner = models.ForeignKey(Partner, null=True, on_delete=models.SET_NULL)
 
-    headers = models.FileField(
-        upload_to=get_message_store, null=True, blank=True)
-    payload = models.FileField(
-        upload_to=get_message_store, null=True, blank=True)
+    headers = models.FileField(upload_to=get_message_store, null=True, blank=True)
+    payload = models.FileField(upload_to=get_message_store, null=True, blank=True)
 
     compressed = models.BooleanField(default=False)
     encrypted = models.BooleanField(default=False)
@@ -368,13 +343,9 @@ class Message(models.Model):
     def as2message(self):
         """ Returns an object of pyas2lib's Message class"""
         if self.direction == 'IN':
-            as2m = As2Message(
-                sender=self.partner.as2partner,
-                receiver=self.organization.as2org)
+            as2m = As2Message(sender=self.partner.as2partner, receiver=self.organization.as2org)
         else:
-            as2m = As2Message(
-                sender=self.organization.as2org,
-                receiver=self.partner.as2partner)
+            as2m = As2Message(sender=self.organization.as2org, receiver=self.partner.as2partner)
 
         as2m.message_id = self.message_id
         as2m.mic = self.mic
@@ -408,8 +379,7 @@ class Message(models.Model):
             response.raise_for_status()
         except requests.exceptions.RequestException as e:
             self.status = 'R'
-            self.detailed_status = \
-                'Failed to send message, error:\n%s' % traceback.format_exc()
+            self.detailed_status = f'Failed to send message, error:\n{traceback.format_exc()}'
             self.save()
             return
 
@@ -423,22 +393,17 @@ class Message(models.Model):
                 # Get the response headers, convert key to lower case
                 # for normalization
                 mdn_headers = dict(
-                    (k.lower().replace('_', '-'), response.headers[k])
-                    for k in response.headers
-                )
+                    (k.lower().replace('_', '-'), response.headers[k]) for k in response.headers)
 
                 # create the mdn content with message-id and content-type
                 # header and response content
-                mdn_content = '%s: %s\n' % (
-                    'message-id', mdn_headers.get('message-id', self.message_id))
-                mdn_content += '%s: %s\n\n' % (
-                    'content-type', mdn_headers['content-type'])
+                mdn_content = f'message-id: {mdn_headers.get("message-id", self.message_id)}\n'
+                mdn_content += f'content-type: {mdn_headers["content-type"]}\n\n'
                 mdn_content = mdn_content.encode('utf-8') + response.content
 
                 # Parse the as2 mdn received
                 as2mdn = As2Mdn()
-                status, detailed_status = as2mdn.parse(
-                    mdn_content, lambda x, y: self.as2message)
+                status, detailed_status = as2mdn.parse(mdn_content, lambda x, y: self.as2message)
 
                 # Update the message status and return the response
                 if status == 'processed':
@@ -446,10 +411,8 @@ class Message(models.Model):
                     run_post_send(self)
                 else:
                     self.status = 'E'
-                    self.detailed_status = \
-                        'Partner failed to process message: %s' % detailed_status
-                Mdn.objects.create_from_as2mdn(
-                    as2mdn=as2mdn, message=self, status='R')
+                    self.detailed_status = f'Partner failed to process message: {detailed_status}'
+                Mdn.objects.create_from_as2mdn(as2mdn=as2mdn, message=self, status='R')
         else:
             # No MDN requested mark message as success and run command
             self.status = 'S'
@@ -475,21 +438,18 @@ class MdnManager(models.Manager):
                 return_url=return_url
             )
         )
-        mdn.headers.save(name='%s.header' % uuid4(),
-                         content=ContentFile(as2mdn.headers_str))
-        mdn.payload.save(name='%s.mdn' % uuid4(),
-                         content=ContentFile(as2mdn.content))
+        filename = f'{uuid4()}.mdn'
+        mdn.headers.save(name=f'{filename}.header', content=ContentFile(as2mdn.headers_str))
+        mdn.payload.save(filename, content=ContentFile(as2mdn.content))
         return mdn
 
 
 def get_mdn_store(instance, filename):
     current_date = timezone.now().strftime('%Y%m%d')
     if instance.status == 'S':
-        target_dir = os.path.join(
-            'messages', '__store', 'mdn', 'sent', current_date)
+        target_dir = os.path.join('messages', '__store', 'mdn', 'sent', current_date)
     else:
-        target_dir = os.path.join(
-            'messages', '__store', 'mdn', 'received', current_date)
+        target_dir = os.path.join('messages', '__store', 'mdn', 'received', current_date)
 
     return '{0}/{1}'.format(target_dir, filename)
 
@@ -509,10 +469,8 @@ class Mdn(models.Model):
     signed = models.BooleanField(default=False)
     return_url = models.URLField(null=True)
 
-    headers = models.FileField(
-        upload_to=get_mdn_store, null=True, blank=True)
-    payload = models.FileField(
-        upload_to=get_mdn_store, null=True, blank=True)
+    headers = models.FileField(upload_to=get_mdn_store, null=True, blank=True)
+    payload = models.FileField(upload_to=get_mdn_store, null=True, blank=True)
 
     objects = MdnManager()
 
@@ -523,17 +481,14 @@ class Mdn(models.Model):
         """ Send the asynchronous MDN to the partner"""
 
         # convert the mdn headers to dictionary
-        headers = HeaderParser().parsestr(
-            self.headers.read().decode())
+        headers = HeaderParser().parsestr(self.headers.read().decode())
 
         # Send the mdn to the partner
         try:
             response = requests.post(
-                self.return_url,
-                headers=dict(headers.items()),
-                data=self.payload.read())
+                self.return_url, headers=dict(headers.items()), data=self.payload.read())
             response.raise_for_status()
-        except requests.exceptions.RequestException as e:
+        except requests.exceptions.RequestException:
             return
 
         # Update the status of the MDN