+++ pyas2/tests/test_advanced.py
@@ -316,7 +316,6 @@ class AdvancedTestCases(TestCase):
         partner = Partner.objects.create(
             name='AS2 Server', as2_name='as2server',
             target_url='http://localhost:8080/pyas2/as2receive',
-            # signature='sha1',
             signature_cert=self.server_crt,
             encryption='tripledes_192_cbc',
             encryption_cert=self.server_crt,
@@ -428,7 +427,10 @@ class AdvancedTestCases(TestCase):
         # Create a file for testing
         outbox_dir = os.path.join(
             settings.DATA_DIR, 'messages', 'as2client', 'outbox', 'as2server')
-        os.makedirs(outbox_dir)
+        try:
+            os.makedirs(outbox_dir)
+        except FileExistsError:
+            pass
         test_file = Path(os.path.join(outbox_dir, 'testmessage.edi'))
         test_file.touch()