+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Release History
 ===============
 
+1.1.0 - 2019-06-13
+------------------
+
+* Use original filename when saving to store and allow search by filename.
+* Bump version of pyas2lib to 1.2.0 to fix issue #5
+* Minimum version of django is now 2.1.9 which fixes issue #8
+* Extract and save certificate information on upload.
+
 1.0.2 - 2019-05-16
 ------------------